requirements.txt
appdirs==1.4.3
dj-database-url==0.4.1
Django==1.10.4
django-datetime-widget==0.9.3
django-mptt==0.8.7
django-tables2==1.3.0
djangorestframework==3.6.2
gunicorn==19.6.0
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
virtualenv==15.1.0
whitenoise==3.2
